refactor: moved template field processing into helpers module
fix: document processing/OCR now works correctly with templates and adheres to style examples
