+++ server/utils/transcription.py
@@ -6,8 +6,9 @@ import logging
 from typing import Dict, List, Union
 from ollama import AsyncClient as AsyncOllamaClient
 from server.database.config import config_manager
+from server.utils.helpers import refine_field_content
 from server.schemas.templates import TemplateField, TemplateResponse
-from server.schemas.grammars import FieldResponse, RefinedResponse
+from server.schemas.grammars import FieldResponse
 
 # Configure logging
 logging.basicConfig(level=logging.INFO)
@@ -224,55 +225,6 @@ def clean_list_spacing(text: str) -> str:
     text = re.sub(r'^\s{2,}', ' ', text)
     return text.strip()
 
-async def refine_field_content(
-    content: Union[str, Dict],
-    field: TemplateField
-) -> Union[str, Dict]:
-    """
-    Refine the content of a single field using style examples and format schema.
-
-    Args:
-        content (Union[str, Dict]): The raw content to refine.
-        field (TemplateField): The field being processed.
-
-    Returns:
-        Union[str, Dict]: The refined content.
-    """
-    try:
-        # If content is already structured (dict), return as is
-        if isinstance(content, dict):
-            return content
-
-        # Get configuration and client
-        config = config_manager.get_config()
-        client = AsyncOllamaClient(host=config["OLLAMA_BASE_URL"])
-        prompts = config_manager.get_prompts_and_options()
-        options = prompts["options"]["general"]
-
-        # Determine format details
-        format_details = _determine_format_details(field, prompts)
-
-        # Build system prompt with style example if available
-        system_prompt = _build_system_prompt(field, format_details, prompts)
-
-        # Execute the model call
-        response = await client.chat(
-            model=config["PRIMARY_MODEL"],
-            messages=[
-                {"role": "system", "content": system_prompt},
-                {"role": "user", "content": content},
-            ],
-            format=format_details["response_format"],
-            options=options
-        )
-
-        # Format the response appropriately
-        return _format_refined_response(response, field, format_details)
-
-    except Exception as e:
-        logger.error(f"Error refining field {field.field_key}: {e}")
-        raise
-
 def _build_patient_context(context: Dict[str, str]) -> str:
     """
     Build patient context string from dictionary.
@@ -334,113 +286,3 @@ def _detect_audio_format(audio_buffer):
         return "recording.m4a", "audio/mp4"
     # Default to WAV if we can't determine
     return "recording.wav", "audio/wav"
-
-def _determine_format_details(field: TemplateField, prompts: dict) -> dict:
-    """Determine response format and format type based on field schema."""
-    format_type = None
-    if field.format_schema and "type" in field.format_schema:
-        format_type = field.format_schema["type"]
-
-        if format_type == "narrative":
-            return {
-                "format_type": "narrative",
-                "response_format": NarrativeResponse.model_json_schema(),
-                "base_prompt": "Format the following content as a cohesive narrative paragraph."
-            }
-
-    # Default to RefinedResponse for non-narrative formats
-    format_guidance = ""
-    if format_type == "numbered":
-        format_guidance = "Format the key points as a numbered list (1., 2., etc.)."
-    elif format_type == "bullet":
-        format_guidance = "Format the key points as a bulleted list (•) prefixes)."
-
-    return {
-        "format_type": format_type,
-        "response_format": RefinedResponse.model_json_schema(),
-        "base_prompt": prompts["prompts"]["refinement"]["system"],
-        "format_guidance": format_guidance
-    }
-
-def _build_system_prompt(field: TemplateField, format_details: dict, prompts: dict) -> str:
-    """Build the system prompt using format guidance and style examples."""
-    # Check if field has style_example and prioritize it
-    if hasattr(field, 'style_example') and field.style_example:
-        return f"""
-        You are an expert medical scribe. Your task is to reformat medical information to precisely match the style example provided, while maintaining clinical accuracy.
-
-        STYLE EXAMPLE:
-        {field.style_example}
-
-        FORMATTING INSTRUCTIONS:
-        1. Analyze the STYLE EXAMPLE and match it EXACTLY, including:
-        - Format elements (bullet style, indentation, paragraph structure)
-        - Sentence structure (fragments vs. complete sentences)
-        - Capitalization and punctuation patterns
-        - Abbreviation conventions and medical terminology style
-        - Tense (past/present) and perspective (first/third person)
-
-        2. IMPORTANT CONSTRAINTS:
-        - Preserve ALL clinical details and values from the original text
-        - Do not add information not present in the input
-        - RETURN JSON IN THE REQUESTED FORMAT
-        - If the style example uses abbreviations like "SNT" or "HSM", use similar appropriate medical abbreviations
-
-        FORMAT THE FOLLOWING MEDICAL INFORMATION:"""
-
-    # If no style example, start with base prompt
-    system_prompt = format_details["base_prompt"]
-
-    # Add format guidance if available
-    if "format_guidance" in format_details and format_details["format_guidance"]:
-        system_prompt += "\n" + format_details["format_guidance"]
-
-    # Apply custom refinement rules if specified and no style example exists
-    if field.refinement_rules:
-        for rule in field.refinement_rules:
-            if rule in prompts["prompts"]["refinement"]:
-                system_prompt = prompts["prompts"]["refinement"][rule]
-                break
-    print(system_prompt)
-    return system_prompt
-
-def _format_refined_response(response: dict, field: TemplateField, format_details: dict) -> str:
-    """Format the model response according to field requirements."""
-    format_type = format_details["format_type"]
-
-    if format_type == "narrative":
-        narrative_response = NarrativeResponse.model_validate_json(response['message']['content'])
-        return narrative_response.narrative
-
-    # Handle non-narrative formats
-    refined_response = RefinedResponse.model_validate_json(response['message']['content'])
-
-    if format_type == "numbered":
-        return _format_numbered_list(refined_response.key_points)
-    elif format_type == "bullet":
-        return _format_bulleted_list(refined_response.key_points, field)
-    else:
-        # No specific formatting required
-        return "\n".join(refined_response.key_points)
-
-def _format_numbered_list(key_points: List[str]) -> str:
-    """Format key points as a numbered list."""
-    formatted_key_points = []
-    for i, point in enumerate(key_points):
-        # Strip any existing numbering
-        cleaned_point = re.sub(r'^\d+\.\s*', '', point.strip())
-        formatted_key_points.append(f"{i+1}. {cleaned_point}")
-    return "\n".join(formatted_key_points)
-
-def _format_bulleted_list(key_points: List[str], field: TemplateField) -> str:
-    """Format key points as a bulleted list."""
-    bullet_char = "•"  # Default bullet character
-    if field.format_schema and "bullet_char" in field.format_schema:
-        bullet_char = field.format_schema["bullet_char"]
-
-    formatted_key_points = []
-    for point in key_points:
-        # Strip any existing bullets
-        cleaned_point = re.sub(r'^[•\-\*]\s*', '', point.strip())
-        formatted_key_points.append(f"{bullet_char} {cleaned_point}")
-    return "\n".join(formatted_key_points)
